added etests for SearchBar
